+++ Runtime/DB/Enemy DB/DamagerTag.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace OcDialogue.DB.Enemy_DB
+{
+    public enum DamagerTag
+    {
+        Weapon1,
+        Weapon2,
+        LeftHand,
+        RightHand,
+        LeftFoot,
+        RightFoot,
+        Head,
+        Tail
+    }
+}